Rewrote level 5 into what it was originally envisioned to be.

+++ src/com/maltelenz/sensortrouble/Level1Screen.java
@@ -39,8 +39,8 @@ public class Level1Screen extends LevelScreen {
     }
 
     @Override
-    float percentDone() {
-        return ((float) (maxTouches - touchesLeft))/maxTouches;
+    double percentDone() {
+        return ((double) (maxTouches - touchesLeft))/maxTouches;
     }
 
     void drawRunningUI() {
